Move enum to service namespace
