+++ templates/home/index/_partial/_missions.html.twig
@@ -17,7 +17,7 @@
             <!--Grid column-->
             <div class="col-lg-4 col-md-6 mission">
                 <h4 class="my-4 font-weight-bold">
-                    {{ nearestMission.state == constant('App\\Enum\\Mission\\MissionStateEnum::ARCHIVED') ? 'Last mission'|trans : 'Nearest mission'|trans }}
+                    {{ nearestMission.state == constant('App\\Service\\Mission\\Enum\\MissionStateEnum::ARCHIVED') ? 'Last mission'|trans : 'Nearest mission'|trans }}
                 </h4>
 
                 {% include '_partial/_mission_widget.html.twig' with {'mission': nearestMission} %}
